types and enums update
