--- types/ParagraphTypeStyle.ts
@@ -1,5 +1,0 @@
-export enum ParagraphTypeStyle {
-    lg = 'body-lg',
-    md = 'body-md',
-    sm = 'body-sm',
-  }
