+++ types/typography/Heading.ts
@@ -1,3 +1,12 @@
+export enum HeadingLevel {
+    h1 = 'h1',
+    h2 = 'h2',
+    h3 = 'h3',
+    h4 = 'h4',
+    h5 = 'h5',
+    h6 = 'h6',
+}
+
 export enum HeadingTypeStyle {
     xl = 'heading-xl',
     lg = 'heading-lg',
@@ -10,4 +19,13 @@ export enum HeadingTypeStyle {
     bodyLg = 'body-lg',
     bodyMd = 'body-md',
     bodySm = 'body-sm',
+}
+
+export interface HeadingProps {
+    typeStyle?: HeadingTypeStyle;
+    level?: HeadingLevel;
+    textColor?: string;
+    className?: string;
+    children?: React.ReactNode;
+    id?: string;
   }
